Combine condense `getFeedItems()` into `getArticles()`
Unfortunately, I can't seem to make gson serialize a (top-level) JSON
array straight to an `ArrayList`. So, I've moved all the retrieving of
articles logic into one method, `getArticles()`.

We grab the articles as an old-fashion `Array` and immediately add them
to the `ArrayList` we plan to return. In this way, we do our best to
remove any references we have to the original array (partially by having
the stack go away on `return`) allowing as much memory as possible to be
recovered _when_ GC is run.